requirements.txt
WMI~=1.5.1
PyYAML~=6.0.2